Implements noaccess check for all routes

+++ handlers/reports.js
@@ -14,6 +14,8 @@ const repackLogs = require('../lib/repack-logs')
 const generateExcelFile = require('../lib/generate-excel-file')
 
 module.exports.generateApplicationsReport = async (request, reply) => {
+  const yar = request.yar
+  const isAdmin = yar.get('isAdmin')
   const userId = request.auth.credentials.data.userId
   const token = generateSystemJwt(userId)
   const url = `${config.LOGS_SERVICE_URL}/logs/search`
@@ -26,37 +28,47 @@ module.exports.generateApplicationsReport = async (request, reply) => {
     }
   }
 
-  axios.defaults.headers.common['Authorization'] = token
+  if (!isAdmin) {
+    reply.redirect('/noaccess')
+  } else {
+    axios.defaults.headers.common['Authorization'] = token
 
-  logger('info', ['reports', 'generateApplicationsReport', 'user', userId, `${request.payload.fromDate} - ${request.payload.toDate}`])
-  const results = await axios.post(url, mongoQuery)
-  const report = results.data.map(repackLogs())
-  const uniqueName = `${uuid.v4()}.xlsx`
-  const filename = `${os.tmpdir()}/${uniqueName}`
+    logger('info', ['reports', 'generateApplicationsReport', 'user', userId, `${request.payload.fromDate} - ${request.payload.toDate}`])
+    const results = await axios.post(url, mongoQuery)
+    const report = results.data.map(repackLogs())
+    const uniqueName = `${uuid.v4()}.xlsx`
+    const filename = `${os.tmpdir()}/${uniqueName}`
 
-  await generateExcelFile({filename: filename, data: report})
-  const excel = fs.readFileSync(filename)
+    await generateExcelFile({filename: filename, data: report})
+    const excel = fs.readFileSync(filename)
 
-  reply(excel)
-    .header('Content-Type', 'application/vnd.openxmlformats-officedocument.spreadsheetml.sheet')
-    .header('Content-Disposition', 'attachment; filename=' + uniqueName)
-    .on('finish', () => {
-      logger('info', ['reports', 'generateApplicationsReport', 'user', userId, `${request.payload.fromDate} - ${request.payload.toDate}`, uniqueName, 'success'])
-      try {
-        fs.unlinkSync(filename)
-        logger('info', ['reports', 'generateApplicationsReport', 'user', userId, `${request.payload.fromDate} - ${request.payload.toDate}`, uniqueName, 'cleanup finished'])
-      } catch (error) {
-        logger('error', ['reports', 'generateApplicationsReport', 'user', userId, `${request.payload.fromDate} - ${request.payload.toDate}`, 'unlink', error])
-      }
-    })
+    reply(excel)
+      .header('Content-Type', 'application/vnd.openxmlformats-officedocument.spreadsheetml.sheet')
+      .header('Content-Disposition', 'attachment; filename=' + uniqueName)
+      .on('finish', () => {
+        logger('info', ['reports', 'generateApplicationsReport', 'user', userId, `${request.payload.fromDate} - ${request.payload.toDate}`, uniqueName, 'success'])
+        try {
+          fs.unlinkSync(filename)
+          logger('info', ['reports', 'generateApplicationsReport', 'user', userId, `${request.payload.fromDate} - ${request.payload.toDate}`, uniqueName, 'cleanup finished'])
+        } catch (error) {
+          logger('error', ['reports', 'generateApplicationsReport', 'user', userId, `${request.payload.fromDate} - ${request.payload.toDate}`, 'unlink', error])
+        }
+      })
+  }
 }
 
 module.exports.getReportFrontpage = async (request, reply) => {
+  const yar = request.yar
+  const isAdmin = yar.get('isAdmin')
   const userId = request.auth.credentials.data.userId
-  const now = new Date()
-  const today = `${now.getFullYear()}-${datePadding(now.getMonth() + 1)}-${datePadding(now.getDate())}`
-  logger('info', ['reports', 'getReportFrontpage', 'user', userId])
-  const viewOptions = createViewOptions({credentials: request.auth.credentials, today: today})
+  if (!isAdmin) {
+    reply.redirect('/noaccess')
+  } else {
+    const now = new Date()
+    const today = `${now.getFullYear()}-${datePadding(now.getMonth() + 1)}-${datePadding(now.getDate())}`
+    logger('info', ['reports', 'getReportFrontpage', 'user', userId])
+    const viewOptions = createViewOptions({credentials: request.auth.credentials, today: today})
 
-  reply.view('reports', viewOptions)
+    reply.view('reports', viewOptions)
+  }
 }
